Automatically adjust resolution
This fixes a problem with periods crossing years #7, but more
importantly it changes the interval used to display tasks based on the
overall time frame (switching from days to weeks and finally months) it
also introduces year quarters for the headers.

This means for large time frames, the displayed accuracy is reduced in
favor of a birdseye view.

This is largely untested and I expect a bunch of off-by-one errors. The
size scaling is probably broken too.

This needs some unit tests.